♻️ Factor out swarpc stuff into... a NPM package, yep lol
see https://github.com/gwennlbh/swarpc
